pyproject.toml
[project]
name = "bilidownloader-cmd"
version = "0.1.0"
description = ""
authors = [
    {name = "mahrat",email = "2658842706@qq.com"}
]
license = {text = "GNUv3"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "bilibili-api-python (>=17.1.4,<18.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "qrcode (>=8.0.0,<9.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "questionary (>=2.1.0,<3.0.0)"
]

[tool.poetry]
packages = [
    { include = "client", from = "src" },
    { include = "server", from = "src" },
    { include = "tools", from = "src" },
    { include = "common", from = "src" },
    { include = "service", from = "src" }
]

[tool.poetry.scripts]
btool-download = "client.cli:cli"
btool-server = "server.server_core:run_server"
btool-genconfig = "tools.create_config:generate_config"

[tool.poetry.group.dev.dependencies]
click = "^8.1.8"
flask = "^3.1.0"
requests = "^2.32.3"
bilibili-api-python = "^17.1.4"
tqdm = "^4.67.1"
rich = "^13.9.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
